fix(cli): address deep code review findings (#1018)
* fix(cli): address deep review findings

- canon check: exit 2 (fail closed) instead of silent no-op exit 0
- oscal validate: exit 2 (fail closed); lazy-import trestle to prevent
  import-time side-effects on every uiao --help
- enforcement/identity/ir: wrap json.loads() in try/except for
  JSONDecodeError and OSError; exit 2 on parse failure
- cql/evidence/ir: replace free-form --format str flags with typed
  Enums (CQLFormat, EvidenceFormat, IRMarkdownFormat, IRTextFormat)
  so invalid values fail loudly with Typer's built-in validation
- init: replace hardcoded /tmp/uiao-quickstart default with
  tempfile.gettempdir() for cross-platform correctness
- app: wire adapter_oscal_app into CLI as 'adapter-oscal' (was dead)
- plantuml: add allow_remote=False guard; remote plantweb fallback
  now requires explicit opt-in to prevent data exfiltration
- ratelimit: add type: ignore[import-untyped] for optional redis extra

Tests: add happy-path + error-path coverage for all changed commands


* fix(mypy): target python_version 3.12; fix auth timeout type; fix init default arg

- pyproject.toml: raise [tool.mypy] python_version from "3.10" to "3.12".
  The CI runner uses Python 3.12 and numpy ≥ 2.4 ships bundled stubs that
  use the PEP 695 `type` statement (3.12+ syntax). Mypy parses stubs before
  per-module overrides apply, so the syntax error was fatal regardless of
  any follow_imports/ignore_errors override.
- pyproject.toml: remove the now-superseded `follow_imports = "skip"` numpy
  override (added in the previous commit). That override does not prevent mypy
  from parsing stub files during module resolution and is no longer needed
  with python_version = "3.12".
- saas/auth.py: fix jwks_verifier timeout: float → int. PyJWKClient expects
  int; the float default was a real type error previously masked by the numpy
  parse failure stopping mypy before it reached auth.py.
- cli/init.py: use None as the --out-dir default (resolve tempfile.gettempdir()
  at call time, not at module import time).

---------

+++ src/uiao/saas/auth.py
@@ -177,7 +177,7 @@ class EntraTokenVerifier:
         )
 
 
-def jwks_verifier(cloud: str = "commercial", *, timeout: float = 5.0) -> SignatureVerifier:
+def jwks_verifier(cloud: str = "commercial", *, timeout: int = 5) -> SignatureVerifier:
     """Return a production JWKS-backed signature verifier.
 
     Lazy-imports ``PyJWT`` (the ``[saas]`` extra). Fetches the issuing
